升级spring boot 版本

+++ feign-client/src/main/java/com/yanglf/feign/UserClientFallback.java
@@ -10,7 +10,7 @@ import com.yanglf.user.model.TbUser;
  */
 public class UserClientFallback implements UserClient {
     @Override
-    public TbUser info(int id) {
+    public TbUser info(Long id) {
         return TbUser.builder()
                 .userName("yanglf")
                 .build();
